+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+## [0.3.2] - 2016-11-25
+
+## Added
+
+- Optimalize code
+- Add getValue() function
+- Add events
+
+## Fixed
+
+- Usage text in README.md
+
 ## [0.3.1] - 2016-11-18
 
 ## Added